% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls-functions.R
\docType{data}
\name{defaultNLSControl}
\alias{defaultNLSControl}
\title{Default control for nonlinear least squares fitting of saturation curves}
\format{An object of class \code{list} of length 5.}
\usage{
defaultNLSControl
}
\description{
This is a wrapper for nls.control. It is a list that can be directly appended.
See nls.control for more details.
}
\keyword{datasets}
