% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls-functions.R
\name{defaultNLSStarts}
\alias{defaultNLSStarts}
\title{Default combinations of starting values for nonlinear least squares fitting
of saturation curves}
\usage{
defaultNLSStarts(maxsig)
}
\arguments{
\item{maxsig}{the maximum number of significant genes in the curve}
}
\value{
a list of lists, each of which serves as a set of starting
parameter estimates for nls
}
\description{
Default combinations of starting values for nonlinear least squares fitting
of saturation curves
}
