% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls-functions.R
\name{fitnls}
\alias{fitnls}
\title{Fit significance curve using nonlinear least squares}
\usage{
fitnls(dat, method = "probit", starts = NULL,
  control = defaultNLSControl)
}
\arguments{
\item{dat}{a data.frame from summary.subsamples, containing columns for
"proportion" and "significant"}

\item{method}{specifies the method used to fit the model. Default is probit (superSeq). "logit" and "smoother" are possible
options but are mainly for use in our manuscript.}

\item{starts}{A list of combinations of starting guesses to try with
\code{nls}. If NULL, uses \code{\link{defaultNLSStarts}}}

\item{control}{control for \code{nls}, by default \link{defaultNLSControl}}
}
\description{
Fit a nonlinear least squares model explaining the number of significant
genes based on the proportion of reads used. Tries each of the starting
combinations of parameters in the given order. If all fail, returns NULL.
}
\seealso{
\code{\link{defaultNLSStarts}}
}
