% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-superSeq.R
\name{plot.superSeq}
\alias{plot.superSeq}
\alias{plot,}
\title{Plotting for superSeq object}
\usage{
\method{plot}{superSeq}(x, ...)
}
\arguments{
\item{x}{superSeq object}

\item{...}{not used}
}
\description{
Plotting for superSeq object
}
\keyword{plot}
