% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-superSeq.R
\name{summary.superSeq}
\alias{summary.superSeq}
\title{Summary function of superSeq object}
\usage{
\method{summary}{superSeq}(object, depth = c(1, 2, 3),
  digits = getOption("digits"), ...)
}
\arguments{
\item{object}{superSeq object}

\item{depth}{Specified read depth proportion to provide predictions in the summary.}

\item{digits}{Number of digits to print.}

\item{...}{Not currently used}
}
\description{
Summary function of superSeq object
}
